Moved `scanned_poem` out of Metroscope pkg and into `/poetry/poem` route
