+++ tests/test_flask/test_poetry_blueprint/test_poem_view.py
@@ -1,5 +1,18 @@
 """Test the poem view using flask's test_client()."""
 
+from run.poetry.views import rhyme_designator
+
+
+def test_rhyme_designator():
+    """Should return a short string designating an index."""
+
+    assert rhyme_designator(0) == "A"
+    assert rhyme_designator(26) == "A1"
+    assert rhyme_designator(52) == "A2"
+    assert rhyme_designator(1) == "B"
+    assert rhyme_designator(27) == "B1"
+    assert rhyme_designator(53) == "B2"
+
 
 def test_poem_no_keyword(client_poems):
     """Check the poem should return a list of available poems."""
